Patch b1 readme reduce (#20)
* requirements.txt updated (tqdm version loosen).

* requirements.txt updated. Removed packages that `datasets` depends on.

* Jury moved under __init__.py. black package added to requirements-dev.txt.

* v0.0.5

* requirements.txt updated. Removed packages that `datasets` depends on.

* Jury moved under __init__.py. black package added to requirements-dev.txt.

* v0.0.5

* reduce_fn added. README.md updated.

* Patch b1:
- README.md updated.
- Utility tests added.
- Test scripts added.
- ci.yml updated.
- InputList renamed as Collator.
- setup.cfg added for code style (flake8).
