+++ requirements-dev.txt
@@ -1,4 +1,12 @@
-black
+datasets>=1.8.0
+fire>=0.4.0
+numpy>=1.21.0
+rouge-score==0.0.4
+tqdm
+# Development
+black==21.7b0
+flake8
 pytest>=6.2.4
+pytest-cov>=2.12.1
 sacrebleu
 bert_score
