+++ setup.cfg
@@ -1,0 +1,5 @@
+[flake8]
+max-line-length = 127
+select = E9,F63,F7,F82
+per-file-ignores = __init__.py: F401
+max-complexity = 10
