+++ tests/run_tests.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+set -xeu
+
+pytest --cov jury --cov-report term-missing --cov-report xml -vvv tests
